feat(PostCard.tsx): Change post card style
