+++ client/src/Common/Types/User/User.tsx
@@ -1,6 +1,7 @@
 type User = {
   id: string;
   username: string;
+  imageUrl?: string;
 };
 
 export default User;
